+++ NEWS.md
@@ -1,3 +1,4 @@
 # greta.distributions 0.0.0.9000
 
+* Added a `greta_distribution_template` function to write a new distribution template
 * Added a `NEWS.md` file to track changes to the package.